LEDとVibrationのデータ取得に失敗していたので修正
